Update ImprovedEases.hx
expoIn ended on 0.999 instead of 1.0
removing the -0.001 fixes the issue.

also added a missing 0 (1000 instead of 100) for impulse which caused it to ending at the incorrect value.

Also ported smoothStepOutIn and smootherStepOutIn from WITF

Note that outElastic might end incorrectly as well and might need fixing, but I haven't tested it.